Add cast flag to only cast spell if target is casting.

+++ src/game/ObjectMgr.cpp
@@ -1777,7 +1777,7 @@ void ObjectMgr::LoadCreatureSpells()
                 uint32 targetParam1    = fields[4 + i * CREATURE_SPELLS_MAX_COLUMNS].GetUInt32();
                 uint32 targetParam2    = fields[5 + i * CREATURE_SPELLS_MAX_COLUMNS].GetUInt32();
 
-                uint8 castFlags        = fields[6 + i * CREATURE_SPELLS_MAX_COLUMNS].GetUInt8();
+                uint16 castFlags        = fields[6 + i * CREATURE_SPELLS_MAX_COLUMNS].GetUInt16();
 
                 // in the database we store timers as seconds
                 // based on screenshot of blizzard creature spells editor
